# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FermentOS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List all inventory items
      parameters:
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/IngredientType"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Add an inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Inventory item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List all equipment
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Equipment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Add a piece of equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentBody"
      responses:
        "201":
          description: Created equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    put:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update a piece of equipment
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEquipmentBody"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete a piece of equipment
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    IngredientType:
      type: string
      enum: [malt, hop, yeast, adjunct, water_agent, other]
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          $ref: "#/components/schemas/IngredientType"
        amount:
          type: number
        unit:
          type: string
        purchasedDate:
          type: string
          format: date
          nullable: true
        expiryDate:
          type: string
          format: date
          nullable: true
        supplier:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, type, amount, unit, createdAt, updatedAt]
    CreateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: "#/components/schemas/IngredientType"
        amount:
          type: number
        unit:
          type: string
        purchasedDate:
          type: string
          format: date
          nullable: true
        expiryDate:
          type: string
          format: date
          nullable: true
        supplier:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, type, amount, unit]
    UpdateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        type:
          $ref: "#/components/schemas/IngredientType"
        amount:
          type: number
        unit:
          type: string
        purchasedDate:
          type: string
          format: date
          nullable: true
        expiryDate:
          type: string
          format: date
          nullable: true
        supplier:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    EquipmentCondition:
      type: string
      enum: [new, good, fair, poor]
    Equipment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        category:
          type: string
        purchasedDate:
          type: string
          format: date
          nullable: true
        purchasePrice:
          type: string
          nullable: true
        condition:
          $ref: "#/components/schemas/EquipmentCondition"
          nullable: true
        serialNumber:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, category, createdAt, updatedAt]
    CreateEquipmentBody:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        model:
          type: string
        category:
          type: string
        purchasedDate:
          type: string
          format: date
        purchasePrice:
          type: string
        condition:
          $ref: "#/components/schemas/EquipmentCondition"
        serialNumber:
          type: string
        notes:
          type: string
      required: [name, category]
    UpdateEquipmentBody:
      type: object
      properties:
        name:
          type: string
        brand:
          type: string
        model:
          type: string
        category:
          type: string
        purchasedDate:
          type: string
          format: date
        purchasePrice:
          type: string
        condition:
          $ref: "#/components/schemas/EquipmentCondition"
        serialNumber:
          type: string
        notes:
          type: string
...